[ref] move planar into dir
Description
So the planar object are seperated from the 3d objects
