+++ rust/src/bin/cornell-box.rs
